+++ src/.internal/core/util/Type.ts
@@ -119,7 +119,7 @@ export function assert(condition: boolean, message: string = "Assertion failed")
 export function toNumber(value: any): number {
 	if (value != null && !isNumber(value)) {
 		let converted = Number(value);
-		if (isNaN(converted) && isString(value) && value != "") {
+		if (isNaN(converted) && isString(value) && value != "" && value.match(/[0-9]+/)) {
 			return toNumber(value.replace(/[^0-9.\-]+/g, ''));
 		}
 		return converted;